+++ CHANGES.md
@@ -10,6 +10,7 @@
 - ngr: Add capability to invoke projects using the `poethepoet` task runner
 - Dependencies: Update to nbdime 4 and pytest-notebook 0.10
 - Add `pueblo.io.to_io` utility function
+- ngr: Improve .NET runner by accepting `--dotnet-version` command-line option
  
 ## 2023-11-06 v0.0.3
 - ngr: Fix `contextlib.chdir` only available on Python 3.11 and newer